% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggperwindow.R
\docType{data}
\name{aggperwindow}
\alias{aggperwindow}
\title{aggperwindow}
\format{
\subsection{\code{aggperwindow}}{

Data frame with 108 observations and 451 columns:
\describe{
\item{output}{data frame with GGIRmatcher window summary output}
}
}
}
\source{
TEMPUS
}
\usage{
aggperwindow
}
\description{
Object usually stored in the GGIRmatcher meta data from part 5 including the
window-level summary for the time-use 24-hour analyses.
}
\keyword{datasets}
