% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseSegment.R
\name{analyseSegment}
\alias{analyseSegment}
\title{Title}
\usage{
analyseSegment(
  ts,
  id,
  splits,
  addFUN,
  addCutpoints = NULL,
  segment_names = NULL,
  lmv = c(30, 100, 430),
  legend
)
}
\arguments{
\item{ts}{}

\item{id}{}

\item{splits}{}

\item{addFUN}{}

\item{addCutpoints}{}

\item{segment_names}{}

\item{lmv}{}

\item{legend}{}
}
\value{
aggregated data frame.
}
\description{
Title
}
