% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_cleaned.R
\docType{data}
\name{cgm_cleaned}
\alias{cgm_cleaned}
\title{Cleaned CGM Example File}
\format{
\subsection{\code{cgm_cleaned}}{

A data frame with 1,339 rows and 2 columns:
\describe{
\item{timestamp}{Character containing time stamp in ISO8601 format}
\item{GLUC}{Numeric. Glucose in mg/dL}
}
}
}
\source{
TEMPUS project
}
\usage{
cgm_cleaned
}
\description{
Cleaned version of the inst/extdata/cgm_raw_sample.csv
}
\keyword{datasets}
