% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_structure.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Creates Folder Structure for GGIRmatcher project}
\usage{
create_folder_structure(
  outputdir,
  GGIR_dirname = "GGIR",
  additional_dirname = "CGM",
  matched_dirname = "GGIRmatcher"
)
}
\arguments{
\item{outputdir}{Character (no default value) with the parent folder in which all the output generated will be stored (ideally an existing empty folder).}

\item{GGIR_dirname}{Character (default = "GGIR") with the name to be given to the folder that will store the GGIR output.}

\item{additional_dirname}{Character (default = "CGM") with the name to be given to the folder that will store the additional sensor output.}

\item{matched_dirname}{Character (default = "GGIRmatcher") with the name to be given to the folder that will store the matched output.}
}
\value{
list including paths to the directories to be used for GGIR, additional sensor, and matched output.
}
\description{
This function creates the folder structure to a standard project that will
use the \code{GGIRmatcher} package. It will create a 3 folders: for the GGIR output,
for the additional sensor output, and for the matched output.
}
