% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\docType{data}
\name{filename}
\alias{filename}
\title{filename as saved in GGIR meta/ms5.outraw folder}
\format{
\subsection{\code{filename}}{

A character vector of length 1:
\describe{
\item{filename}{String including the filename}
}
}
}
\source{
GGIR
}
\usage{
filename
}
\description{
Object usually stored in the GGIR meta data from part 5
}
\keyword{datasets}
