% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_window_number.R
\name{fix_window_number}
\alias{fix_window_number}
\title{Make window number consistent between GGIR output and GGIRmatcher output
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
fix_window_number(WS, ggir_output, ts = NULL)
}
\arguments{
\item{WS}{data frame with output from GGIRmatcher}

\item{ggir_output}{data frame with output from GGIR part 5}

\item{ts}{Matched time series with \link{match_time_series}}
}
\value{
Merged output with GGIR and GGIRmatcher window-level estimates.
}
\description{
Fix window numbers to make them consistent between GGIR output and
GGIRmatcher output. This is specially relevant when additional devices has
started the recording in different dates than the accelerometers. Also relevant
for OO windows as calendar dates in GGIR output may differ from the start_date
in GGIRmatcher depending on whether sleep onset is before or after midnight.
}
