% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_window_report.R
\name{full_window_report}
\alias{full_window_report}
\title{Combines Window-Level Output
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
full_window_report(
  GGIRmatcher_outputdir,
  GGIR_outputdir = NULL,
  includecrit_day_spt = NULL,
  includecrit_day = NULL,
  includecrit_spt = NULL,
  additional_sf_seconds = NULL,
  minimum_WW_OO_length = 0,
  minimum_segments_length = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{GGIRmatcher_outputdir}{Pathname to folder to store the full window-level reports.}

\item{GGIR_outputdir}{Path to folder where the GGIR output is stored.}

\item{includecrit_day_spt}{Criteria to include data from the additional sensor in
the full window as a proportion of recordings available from the total expected
recordings in window (i.e., 0.7 for 70\% of the recordings available).}

\item{includecrit_day}{Criteria to include data from the additional sensor in
the awake hours window as a proportion of recordings available from the total expected
recordings in window (i.e., 0.7 for 70\% of the recordings available).}

\item{includecrit_spt}{Criteria to include data from the additional sensor in
the sleep period time window as a proportion of recordings available from the total expected
recordings in window (i.e., 0.7 for 70\% of the recordings available).}

\item{additional_sf_seconds}{Expected sampling frequency of the additional sensor in seconds.
For example, \code{additional_sf_seconds = 900}, would assume that the additional sensor collects
data every 15 minutes, and therefore a maximum of 96 recordings are expected in a 24-hour window.}

\item{minimum_WW_OO_length}{Minimum window length required to consider a valid window in hours for the WW and the OO window types.}

\item{minimum_segments_length}{Minimum window length required to consider a valid window for the segments of the day in hours.}

\item{verbose}{Logical indicating whether to print progress messages in the console.}

\item{...}{GGIR output parameters (for example, sep_reports and dec_reports can be used).}
}
\value{
does not return anything. It stores reports in the results folder.
}
\description{
Takes the window-level output stored in the meta/ms5.out folder and combines
them into full window-level report that are stored in csv format in the
results folder.
}
\references{
GGIR R Package \url{https://github.com/wadpac/GGIR}
}
