% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvaliddaysindices.R
\name{getValidDayIndices}
\alias{getValidDayIndices}
\title{Get Indices for Valid Windows
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
getValidDayIndices(
  x,
  includecrit_day_spt = NULL,
  includecrit_day = NULL,
  includecrit_spt = NULL,
  additional_sf_seconds = NULL,
  min_window_length = NULL,
  params_output,
  params_cleaning
)
}
\arguments{
\item{x}{Window-level data frame.}

\item{includecrit_day_spt}{Criteria to include data from the additional sensor in
the full window as a proportion of recordings available from the total expected
recordings in window (i.e., 0.7 for 70\% of the recordings available).}

\item{includecrit_day}{Criteria to include data from the additional sensor in
the awake hours window as a proportion of recordings available from the total expected
recordings in window (i.e., 0.7 for 70\% of the recordings available).}

\item{includecrit_spt}{Criteria to include data from the additional sensor in
the sleep period time window as a proportion of recordings available from the total expected
recordings in window (i.e., 0.7 for 70\% of the recordings available).}

\item{additional_sf_seconds}{Expected sampling frequency of the additional sensor in seconds.
For example, \code{additional_sf_seconds = 900}, would assume that the additional sensor collects
data every 15 minutes, and therefore a maximum of 96 recordings are expected in a 24-hour window.}

\item{min_window_length}{Minimum window length required to consider a valid window in hours.}

\item{params_output}{See help documentation in the \code{GGIR} function.}

\item{params_cleaning}{See help documentation in the \code{GGIR} function.}
}
\value{
Matrix with indices for valid windows for GGIR and for additional sensor output
}
\description{
From the window-level full dataset, it extracts the indices (rows) for the
windows that are considered valid based on the criteria specified by the user.#'
}
