% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGIRmatcher-package.R
\docType{package}
\name{GGIRmatcher-package}
\alias{GGIRmatcher}
\alias{GGIRmatcher-package}
\title{GGIRmatcher: Matches GGIR Output With Additional Sensors}
\description{
GGIRmatcher aims to match time series GGIR part 5 output with time series extracted from an additional sensor.
}
\author{
\strong{Maintainer}: Jairo H. Migueles \email{jairo@jhmigueles.com} (\href{https://orcid.org/0000-0003-0366-6935}{ORCID})

Other contributors:
\itemize{
  \item PROFITH research group [funder]
  \item Spanish Ministry of Culture and Sport [funder]
  \item Consejo Superior de Deportes (EXP_77437) [funder]
  \item University of Granada [funder]
}

}
\keyword{internal}
