% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGIRversion.R
\docType{data}
\name{GGIRversion}
\alias{GGIRversion}
\title{GGIRversion}
\format{
\subsection{\code{GGIRversion}}{

A character vector of length 1:
\describe{
\item{GGIRversion}{String including the GGIRversion}
}
}
}
\source{
GGIR
}
\usage{
GGIRversion
}
\description{
Object usually stored in the GGIR meta data from part 5
}
\keyword{datasets}
