% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_timestamp.R
\docType{data}
\name{last_timestamp}
\alias{last_timestamp}
\title{last_timestamp}
\format{
\subsection{\code{last_timestamp}}{

A character vector of length 1:
\describe{
\item{last_timestamp}{String including the last timestamp in the sample data}
}
}
}
\source{
GGIR
}
\usage{
last_timestamp
}
\description{
Object usually stored in the GGIR meta data from part 5
}
\keyword{datasets}
