% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_time_series.R
\name{match_time_series}
\alias{match_time_series}
\title{Matches GGIR Time Series with Other Additional Time Series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
match_time_series(
  GGIR_output_dir,
  additional_ts_dir,
  outputdir,
  add_metric_name = NA,
  idloc = "_",
  tz = Sys.timezone(),
  overwrite = F,
  verbose = T
)
}
\arguments{
\item{GGIR_output_dir}{GGIR output directory (e.g., "output_myproject/"). This
output directory is expected to contain only one single time-series subdirectory.
GGIR generates a subdirectory for each combination of cut-points provided. If more
than one is found, only the first folder (in alphabetical order) will be used for now.
Also, files are expected to be stored in RData format. If you are interested in
extending this feature to work with multiple thresholdn combinations, or with
other data formats, please contact the package maintainer.}

\item{additional_ts_dir}{Additional output directory (e.g., "my_other_sensor_timeseries/").
This is expected to contain an RData file per each recording. RData files should
contain a data frame (and only a data frame) with two columns: timestamp and the
metric to include.}

\item{outputdir}{Character with path to output directory to store the matched
time series and MM, WW, and OO reports.}

\item{add_metric_name}{Character indicating the name to be given to the new added metric.}

\item{idloc}{Character indicating how to find the id within the filename in the
time series. This is expected to match between GGIR file names and additional
output directory file names. For example, if a GGIR time series file is named
"ID001_myproject_hipdata.RData" and \code{idloc = "_"}, then the extracted
ID would be "ID001".}

\item{tz}{Character indicating the time zone to read the timestamps.}

\item{overwrite}{Whether to overwrite the previously generated time series.}

\item{verbose}{Logical indicating whether to print progress messages in console}
}
\value{
Does not return any value, it stores RData files with merged time series
}
\description{
Matches the GGIR exported time series in part 5 with other additional
user-provided time series at second level.
}
