% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat.R
\docType{data}
\name{mdat}
\alias{mdat}
\title{mdat}
\format{
\subsection{\code{mdat}}{

Data frame with 20910 observations and 12 columns:
\describe{
\item{mdat}{data frame with GGIR part 5 output}
}
}
}
\source{
GGIR
}
\usage{
mdat
}
\description{
Object usually stored in the GGIR meta data from part 5 including the
epoch-level time series used for the time-use 24-hour analyses.
}
\keyword{datasets}
