% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{data}
\name{output}
\alias{output}
\title{output}
\format{
\subsection{\code{output}}{

Data frame with 108 observations and 133 columns:
\describe{
\item{output}{data frame with GGIR part 5 output}
}
}
}
\source{
GGIR
}
\usage{
output
}
\description{
Object usually stored in the GGIR meta data from part 5 including the
window-level summary for the time-use 24-hour analyses.
}
\keyword{datasets}
