% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_expansion_log.R
\docType{data}
\name{tail_expansion_log}
\alias{tail_expansion_log}
\title{tail_expansion_log}
\format{
\subsection{\code{tail_expansion_log}}{

A character vector of length 1:
\describe{
\item{tail_expansion_log}{Numeric with the time expanded in GGIR}
}
}
}
\source{
GGIR
}
\usage{
tail_expansion_log
}
\description{
Object usually stored in the GGIR meta data from part 5
}
\keyword{datasets}
