% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_per_window.R
\name{aggregate_per_window}
\alias{aggregate_per_window}
\title{Aggregates Time Series Per Window}
\usage{
aggregate_per_window(
  tspath,
  outputdir,
  GGIR_output_dir,
  FUNs = list(n = function(x) sum(!is.na(x)), mean = mean),
  qwindow = NULL,
  qwindow_names = NULL,
  overwrite = FALSE,
  tz = Sys.timezone(),
  verbose = TRUE
)
}
\arguments{
\item{tspath}{path to RData files containing the matched time series as done with
\code{match_time_series}}

\item{outputdir}{}

\item{GGIR_output_dir}{segments. Only integers allowed for now.}

\item{FUNs}{}

\item{qwindow}{Numeric vector with the hours in the day for which extract}

\item{qwindow_names}{}

\item{overwrite}{}

\item{tz}{Character indicating the time zone to read the timestamps.}

\item{verbose}{}
}
\value{
aggregated data frame.
}
\description{
Aggregates time series per window of interest, including midnight-to-midnight,
wakingup-to-wakingup, sleeponset-to-sleeponset, and user-defined segments
based on parameter \code{hour_splits}.
}
