% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_per_window.R
\name{aggregate_per_window}
\alias{aggregate_per_window}
\title{Aggregates Time Series Per Window
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
aggregate_per_window(
  GGIRmatcher_outputdir,
  GGIR_outputdir,
  FUNs = list(n = function(x) sum(!is.na(x)), mean = mean),
  qwindow = NULL,
  qwindow_names = NULL,
  overwrite = FALSE,
  tz = Sys.timezone(),
  verbose = TRUE
)
}
\arguments{
\item{GGIRmatcher_outputdir}{Path to the directory where the output of GGIRmatcher should be stored.}

\item{GGIR_outputdir}{Path to folder where the GGIR output is stored.}

\item{FUNs}{Named list of functions to be applied to the additional sensor data.}

\item{qwindow}{Same as in GGIR, but it only supports integer vector for now.}

\item{qwindow_names}{Character vector with names to apply to the qwindows.}

\item{overwrite}{Logical indicating whether to overwrite previously-derived output.}

\item{tz}{Character indicating the time zone to read the timestamps.}

\item{verbose}{Logical indicating whether to print progress messages in the console.}
}
\value{
aggregated data frame.
}
\description{
Aggregates time series per window of interest, including midnight-to-midnight,
wakingup-to-wakingup, sleeponset-to-sleeponset, and user-defined segments
based on parameter \code{qwindow}.
}
