% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseSegment.R
\name{analyseSegment}
\alias{analyseSegment}
\title{Title}
\usage{
analyseSegment(
  ts,
  id,
  splits,
  FUNs,
  addCutpoints = NULL,
  qwindow_names = NULL,
  lmv = c(30, 100, 430),
  window_type = "segments",
  legend = NULL
)
}
\arguments{
\item{ts}{}

\item{id}{}

\item{splits}{}

\item{FUNs}{}

\item{addCutpoints}{}

\item{qwindow_names}{}

\item{lmv}{}

\item{window_type}{}

\item{legend}{}
}
\value{
aggregated data frame.
}
\description{
Title
}
