% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Loads RData File with Assigned Name}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{Path to the RData object to load.}
}
\value{
The content of the RData file under the assigned name
}
\description{
Loads RData file with single object and assigns the user-defined name to
the loaded object.
}
\examples{
\dontrun{
x = load("myfile.RData")
exists("x")
}

}
